auto-20150107-023335

+++ dkh/arc.c
@@ -56,7 +56,7 @@ struct workload
   char *host;
   int disk_num;
   int type;
-  double offset;
+  long long offset;
   double size;
   long respone; //respones time
 
@@ -626,7 +626,7 @@ int read_column(struct workload *wl, char *buf)
       // case 2 : wl->host = strdup(tmp); break;
       // case 3 : wl->disk_num = atoi(tmp); break;
       case 4 : (wl->type) = (strcmp(tmp, "Read") == 0) ? READ : WRITE; break;
-      case 5 : wl->offset = atol(tmp); break;
+      case 5 : wl->offset = atoll(tmp); break;
       case 6 : wl->size = atol(tmp); break;
       // case 7 : wl->respone = atol(tmp); break;
     }
